Split Dashboard screen into MainDashboard and StocksList screens.
