+++ pwa/src/pages/StocksList/components/DependentView.tsx
